implementing step 9: Add messages to the on-device index
4. whenever there is a new message, index it, When the viewHolder it is assigned to
is populated in the same FirebaseRecyclerAdapter

This serves the purpose: every time a user sends a message from their device, the message
will be added to the on-device index to be ***discoverable through the Google App*** =
To get public and personal content from the app into users’ search results in the Google App,
i.e. public content will be included in Google Search results and personal content will be included in Google's on-device In Apps search.