implementing step 9: Add messages to the on-device index
5.  Logging user actions on Personal content, e.g. viewing messages,
should be marked with the "upload" attribute set to false in the Metadata object of the Action.
This way we make sure that the user action activity on Personal content remains exclusively on the device, and will Not be uploaded to Google servers.
-> add method getMessageViewAction to prepare a customized Action for that purpose.

This serves the purpose: every time a user sends a message from their device, the message
will be added to the on-device index to be ***discoverable through the Google App*** =
To get public and personal content from the app into users’ search results in the Google App,
i.e. public content will be included in Google Search results and personal content will be included in Google's on-device In Apps search.