implementing step 9: Add messages to the on-device index (aka write app's content to the on-Device index for discoverability in the Google App (set to Personal/ In App search))
6. Finally, add the FirebaseUserActions.getInstance().end(...) method to indicate the end of logging a user action, right after you've added the new message to the index.
 Passing getMessageViewAction() to attain the prepared customized View-Action.

This serves the purpose: every time a user sends a message from their device, the message
will be added to the on-device index to be ***discoverable through the Google App*** =
To get public and personal content from the app into users’ search results in the Google App,
i.e. public content will be included in Google Search results and personal content will be included in Google's on-device In Apps search.

How to test the implementation:
1. start the App and send a new message with the text Athletes
2. Open the Google App, switch to In Apps (aka Personal) and search for Athletes
3. The message will be found in the Google App
4. Tap on the Result -> my App opens at the place where the message with text Athletes is.