Implementing step 10: Receive Re-engagement Notifications
6. add the service declarations for MyFirebaseMessagingService and MyFirebaseInstanceIdService to the Manifest (as children of the application-element).
They are provided with the actions to which they respectivley should respond and made secure by including the android:exported attribute set to false.

About Service declaration:
a) The android:name attribute is the only required attribute—it specifies the class name of the service.
b) Caution: To ensure your app is secure, always use an Explicit intent when starting a Service and do not declare intent filters for your services. Using an implicit intent to start a service is a security hazard because you cannot be certain what service will respond to the intent, and the user cannot see which service starts.
c) But: you can make sure that your service is available Only to your app (as intended) by setting the
android:exported attribute to false. This prevents other apps from starting your service - even When they are using an Explicit intent.